+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Changed
 
 - concord-server: merge existing process values with form values;
+- concord-server: fix update for LDAP group mappings;
 - upgrade BPM version to 0.2.1.
-
 
 
 ## [0.3.1] - 2017-04-02